+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.4.0
+
+* renamed `ReplicaSetPool` to MultiNodePool(breaking)
+* Made `metrix` instrumentation configurable(breaking)
+
 ## 0.3.0
 
 * Reworked instrumentation(breaking if used)